backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
anthropic==0.37.1
pydantic==2.9.2
python-dotenv==1.0.1
slowapi==0.1.9
